requirements.txt
django==4.2
djangorestframework==3.15.1
django-environ==0.11.2
drf-spectacular==0.27.2
